SCOANSI Addition to reverse effect of CSI=F
The effects of both CSI=F and CSI=G are both reversed by this SCOANSI sequence.

The ECMA-48 reference for this is "Framed" which should be canceled by
"51", "54" or "0". I suggest this sequence is always used as:

   CSI 51 ; 0 m

The SCOANSI definition is that it resets all special colour pairs to
their defaults.