Integrated puttyclip patch
Source: <http://www.dmst.aueb.gr/dds/sw/windows/puttyclip/>
Credit: Diomidis Spinellis

puttyclip is a clever patch which adds a pseudo-printer option for ansi
printing which, instead of printing, copies the printed info to the
windows clipboard.  Ansi printing is accomplished by sending a starting
escape sequence, `\e[5i`, followed by text to print, followed by the
stop sequence, `\e[4i`. Putty supports ansi printing natively, when you
select a local printer to send the text to. This patch harnesses that
infrastructure to copy the "printed" text to the local clipboard. When
combined with a simple shell script, such as [winclip][], you can pipe
text to your local clipboard from the host.

This patch seems to not have been updated since the putty 0.53b era, so
it took some doing to get it modernized somewhat. Seems to be functional
but not exactly pretty (yet) and certainly not cross platform.

[winclip]: <http://www.dmst.aueb.gr/dds/sw/windows/puttyclip/winclip>