Add palette manipulation sequences OSC 4 and OSC 10-12
These allow setting the ANSI 256 color palette with the sequence

    ^[]4;n;rgb:rr/gg/bb

which sets color n to the hex color #rrggbb. Similarly:

* OSC 10 sets foreground color
* OSC 11 sets background color
* OSC 12 sets the cursor (background) color

E.g.

    ^[]10;rgb:ff/ff/ff

sets the foreground color to white (#ffffff). This format seems to be
supported by a number of modern terminal emulators (incl. xterm).
Implemented to be compliant with the [Base16] theme approach.

[Base16]: https://github.com/chriskempson/base16

TODO: The palette setting code of PuTTY configuration special cased
the setting of colors >16 to allow option processing to set bg and fg
colors etc. E.g. the option color21 means the cursor color (which is
actually at 261). Need to deal with this.